pyproject.toml
[project]
name = "jerry-bot"
version = "0.1.0"
description = "A (not so) simple Discord bot"
authors = [
    { name = "squid1127", email = "120748767+squid1127@users.noreply.github.com." },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11, <4.0"

dependencies = [
    "aiohttp",                                                           # Plugin: Commands
    "beautifulsoup4",                                                    # Plugin: Commands
    "aiofiles",                                                          # Plugin: Music
    "mutagen",                                                          # Plugin: Music
    "discord.py[voice]",                                               # Voice support
    "squid-core @ git+https://github.com/squid1127/squid-core.git@main",
    "google-genai", # Plugin: Gemini
    "jinja2", # Plugin: Auto Reply - Text Template
    "asteval", # Plugin: Auto Reply - Math Evaluation
]

[project.scripts]
jerry = "jerry_bot.run:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

